move revision interface to entities dir
